Removed support for abs/absorption and added error message
Dimensionless quantaties such as "abs"/"absorbance" are now removed as units. The Type attribute given in the Replicate class is now exclusively reserved for such quantities, while the unit itself remains as dimensionless.

Furthermore, an error message was introduced to notify that a currently not supported unit was given. In addition, all supported prefixes and units will be displayed.